move audio buffer logic to media-utils
