+++ packages/core/src/audio/AudioForRendering.tsx
@@ -15,7 +15,6 @@ import {useAbsoluteCurrentFrame, useCurrentFrame} from '../use-frame';
 import {evaluateVolume} from '../volume-prop';
 import {RemotionAudioProps} from './props';
 import {useFrameForVolumeProp} from './use-audio-frame';
-import {uint8ToBase64} from './audio-url-helpers';
 
 const AudioForRenderingRefForwardingFunction: React.ForwardRefRenderFunction<
 	HTMLAudioElement,
@@ -28,7 +27,6 @@ const AudioForRenderingRefForwardingFunction: React.ForwardRefRenderFunction<
 	const frame = useCurrentFrame();
 	const sequenceContext = useContext(SequenceContext);
 	const {registerAsset, unregisterAsset} = useContext(CompositionManager);
-	const { audioBuffer } = props;
 
 	// Generate a string that's as unique as possible for this asset
 	// but at the same time the same on all threads
@@ -39,15 +37,6 @@ const AudioForRenderingRefForwardingFunction: React.ForwardRefRenderFunction<
 			}-${sequenceContext?.durationInFrames}-muted:${props.muted}`,
 		[props.muted, props.src, sequenceContext]
 	);
-
-	const audioSrc = useMemo(() => {
-		if (audioBuffer) {
-			const arrayBufferAsBase64 = uint8ToBase64(audioBuffer);
-			return 'data:audio/wav;base64,' + arrayBufferAsBase64;
-		}
-
-		return props.src;
-	}, [audioBuffer, props.src])
 
 	const {volume: volumeProp, playbackRate, ...nativeProps} = props;
 
@@ -62,7 +51,7 @@ const AudioForRenderingRefForwardingFunction: React.ForwardRefRenderFunction<
 	});
 
 	useEffect(() => {
-		if (!audioSrc) {
+		if (!props.src) {
 			throw new Error('No src passed');
 		}
 
@@ -72,20 +61,18 @@ const AudioForRenderingRefForwardingFunction: React.ForwardRefRenderFunction<
 
 		registerAsset({
 			type: 'audio',
-			src: getAbsoluteSrc(audioSrc),
+			src: getAbsoluteSrc(props.src),
 			id,
 			frame: absoluteFrame,
 			volume,
-			isRemote: props.audioBuffer
-				? true
-				: isRemoteAsset(getAbsoluteSrc(audioSrc)),
+			isRemote: isRemoteAsset(getAbsoluteSrc(props.src), true),
 			mediaFrame: frame,
 			playbackRate: props.playbackRate ?? 1,
 		});
 		return () => unregisterAsset(id);
 	}, [
 		props.muted,
-		audioSrc,
+		props.src,
 		registerAsset,
 		absoluteFrame,
 		id,
@@ -100,8 +87,8 @@ const AudioForRenderingRefForwardingFunction: React.ForwardRefRenderFunction<
 
 	const propsToPass = {
 		...nativeProps,
-		src: audioSrc
-	}
+		src: props.src,
+	};
 
 	return <audio ref={audioRef} {...propsToPass} />;
 };
