+++ packages/core/src/is-remote-asset.ts
@@ -1,2 +1,5 @@
-export const isRemoteAsset = (asset: string) =>
-	!asset.startsWith(window.location.origin) && !asset.startsWith('data');
+// TODO: Can be removed in 3.0 branch
+
+export const isRemoteAsset = (asset: string, dataURLIsRemoteObject: boolean) =>
+	!asset.startsWith(window.location.origin) ||
+	(asset.startsWith('data') && dataURLIsRemoteObject);
